+++ web/src/constants/auth.ts
@@ -1,0 +1,12 @@
+export const AUTH_CHECK = 'auth/CHECK';
+export const AUTH_CHECK_INITIATED = 'auth/CHECK_INITIATED';
+export const AUTH_CHECK_AUTHORIZED = 'auth/IS_AUTHED';
+export const AUTH_CHECK_UNAUTH = 'auth/NOT_AUTHED';
+export const AUTH_CHECK_FAILED = 'auth/CHECK_FAILED';
+
+export const LOGIN_INITIATED = 'login/INITIATED';
+export const LOGIN_SUCCESSFUL = 'login/SUCCESSFUL';
+
+export const LOGOUT_INITIATED = 'logout/INITIATED';
+export const LOGOUT_SUCCESSFUL = 'logout/SUCCESSFUL';
+export const LOGOUT_FAILED = 'logout/FAILED';
