--- web/src/constants/index.ts
@@ -1,4 +1,0 @@
-export const INCREMENT_REQUESTED = 'counter/INCREMENT_REQUESTED';
-export const INCREMENT = 'counter/INCREMENT';
-export const DECREMENT_REQUESTED = 'counter/DECREMENT_REQUESTED';
-export const DECREMENT = 'counter/DECREMENT';
