+++ web/src/reducers/index.ts
@@ -1,12 +1,11 @@
 import { combineReducers } from 'redux';
-import counter from './counter';
 import auth, { State as AuthState } from './auth';
 
+// A type that represents the entire app state
 export interface AppState {
   auth: AuthState;
 }
 
 export default combineReducers({
-  counter,
   auth,
 });
